Only suppress auto-prompt with GPC if SaleOfInfo is the only displayed tracking purpose
